gradio_app/requirements.txt
gradio
dlib
opencv-python
numpy
Pillow